Added an admin page for swapping faces.
go to url/admin_cube, click on pixel mapping mode.  It will
walk through each pixel and highlight it white, and you click
on the face it appears on the actual cube, and it swaps them.
Then that face turns red, the next one turns white, so and on so forth.